scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.11.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "rasterio>=1.5.0",
]

[tool.uv.workspace]
members = [
    "standalone_nc2tile",
]
